requirements.txt
Flask==2.2.3
python-dotenv>=0.19.2
requests>=2.27.1
flask-cors>=3.0.10
yt-dlp
celery==5.2.7
redis
schema==0.7.5
authlib>=1.0